Refactor visualization classes into a dedicated pallet data module
- Moved enum and data classes from visualization_3d.py to new utils/pallet_data.py
- Added type hints to class attributes
- Simplified visualization_3d.py by importing classes from pallet_data
- Improved code organization and modularity
